decompose elements from main container
