+++ src/components/FirstSection/FirstSection.module.css
@@ -1,0 +1,47 @@
+.firstSection {
+  display: inline-flex;
+  color: #000;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  gap: 10em;
+  margin-top: 5%;
+  margin-bottom: 15%;
+}
+
+.firstSectionTitle {
+  color: #000;
+  font-family: "MulishRegular";
+  font-size: 48px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 56px;
+}
+
+.firstSectionSubtitle {
+  color: #7d7987;
+  font-family: "MulishExtraLight";
+  font-size: 21px;
+  font-style: normal;
+  font-weight: 300;
+  line-height: 32px;
+}
+
+.firstSectionBtn {
+  margin-top: 50px;
+  width: 200px;
+  height: 56px;
+  border-radius: 55px;
+  background: #458ff6;
+  font-family: "MulishRegular";
+  border: 0px;
+  color: #fff;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 54px;
+}
+
+.firstSectionImage {
+  width: 60%;
+}
