+++ src/components/FiveSection/FiveSection.module.css
@@ -1,0 +1,28 @@
+.fiveSection {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  margin-bottom: 15%;
+}
+
+.backgroundImage {
+  position: absolute;
+  object-fit: cover;
+  z-index: -1;
+  overflow: hidden;
+}
+
+.fiveSectionBGMatrix1 {
+  margin-top: -1%;
+  right: 5%;
+  z-index: 0;
+  width: 11%;
+}
+
+.fiveSectionBGMatrix2 {
+  left: 5%;
+  margin-top: 20%;
+  z-index: -1;
+  width: 11%;
+}
