+++ src/components/FiveSection/Slider/Slider.module.css
@@ -1,0 +1,179 @@
+.sliderDiv {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+}
+
+.slide {
+  padding: 64px 118px;
+  text-align: center;
+  height: 350px;
+  width: 65%;
+  border-radius: 24px;
+  background: linear-gradient(208deg, #67c3f3 9.05%, #5a98f2 76.74%);
+}
+
+.sliderTitle {
+  color: #fff;
+  font-family: "MulishRegular";
+  font-size: 36px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 48px; /* 133.333% */
+  margin-block: 0;
+}
+
+.delimeterSlider {
+  width: 100%;
+  margin: 2% 0%;
+  width: 56px;
+  height: 2px;
+  flex-shrink: 0;
+  border-radius: 5px;
+  background: #fff;
+  display: inline-block;
+  margin-block: 0;
+}
+
+.slideContent {
+  margin: 8% 0px;
+  justify-content: center;
+  text-align: left;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
+
+.userInfo {
+  width: 50%;
+  display: flex;
+}
+
+.avatarContainer {
+  width: 145px;
+  height: 145px;
+  overflow: hidden;
+  border: 4px solid white;
+  border-radius: 50%;
+  margin-right: 2em;
+}
+
+.userAvatar {
+  width: 105%;
+  height: 107%;
+  object-fit: cover;
+  object-position: center center;
+  display: block;
+}
+
+.userTextInfo {
+  align-self: center;
+}
+
+.userName {
+  margin-block: 0px;
+  color: #fff;
+  font-family: "MulishRegular";
+  font-size: 22px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 35px;
+}
+
+.userPosition {
+  margin-block: 0px;
+  color: rgba(255, 255, 255, 0.85);
+  font-family: "MulishExtraLight";
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 30px;
+}
+
+.slideReview {
+  width: 50%;
+  color: rgba(255, 255, 255, 0.9);
+  font-family: "MulishExtraLight";
+  font-size: 19px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 30px;
+}
+
+.pager {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 3%;
+  width: 25%;
+  justify-content: space-between;
+}
+
+.dotsList {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 35%;
+}
+
+.pageDot {
+  width: 10px;
+  height: 10px;
+  border-radius: 50%;
+  background-color: rgba(91, 155, 243, 0.4);
+  margin: 0 5px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.pageDot.active {
+  background-color: #5b9bf3;
+}
+
+.arrowPager {
+  background-color: white;
+  border: 0;
+}
+
+.arrowCornerPosition {
+  background-color: white;
+  border: 0;
+  color: rgba(91, 155, 243, 0.4);
+}
+
+.thirdSectionBtn {
+  margin-top: 50px;
+  width: 200px;
+  height: 56px;
+  border-radius: 55px;
+  background: #fff;
+  color: #458ff6;
+  font-family: "MulishRegular";
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 54px;
+  border: 1.4px solid #458ff6;
+}
+
+.backgroundImage {
+  position: absolute;
+  object-fit: cover;
+  z-index: -1;
+  overflow: hidden;
+}
+
+.fiveSectionBGMatrix1 {
+  margin-top: -1%;
+  right: 5%;
+  z-index: 0;
+  width: 11%;
+}
+
+.fiveSectionBGMatrix2 {
+  left: 5%;
+  margin-top: 20%;
+  z-index: -1;
+  width: 11%;
+}
