+++ src/components/FiveSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FiveSection";
