+++ src/components/FourSection/FoursSection.module.css
@@ -1,0 +1,66 @@
+.fourthSection {
+  display: inline-flex;
+  color: #000;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  margin-bottom: 15%;
+}
+
+.fourthSectionContent {
+  width: 50%;
+  margin-right: 10%;
+  margin-left: 8%;
+}
+
+.delimeter {
+  margin: 2% 0%;
+  width: 56px;
+  height: 2px;
+  flex-shrink: 0;
+  border-radius: 5px;
+  background: #000;
+}
+
+.fourthSectionTitle {
+  color: #000;
+  font-family: "MulishRegular";
+  font-size: 48px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 56px;
+}
+
+.fourthSectionSubtitle {
+  color: #7d7987;
+  font-family: "MulishExtraLight";
+  font-size: 21px;
+  font-style: normal;
+  font-weight: 300;
+  line-height: 32px;
+}
+
+.fourthSectionBtn {
+  margin-top: 50px;
+  width: 200px;
+  height: 56px;
+  border-radius: 55px;
+  background: #fff;
+  color: #458ff6;
+  font-family: Mulish;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 54px;
+  border: 1.4px solid #458ff6;
+}
+
+.downArrowInBtn {
+  width: 20px;
+  height: 12px;
+  flex-shrink: 0;
+}
+
+.fourthSectionImage {
+  width: 60%;
+}
