+++ src/components/SecondSection/ServiceCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ServiceCard";
