+++ src/components/SixSection/ArticleCard/ArticleCard.module.css
@@ -1,0 +1,87 @@
+.cardGrid {
+  display: grid;
+  grid-template-columns: repeat(3, minmax(20em, 1fr));
+  gap: 40px;
+  margin-bottom: 2%;
+}
+
+.articleCard {
+  width: 320px;
+  height: 500px;
+  border-radius: 20px;
+  background: #fff;
+  box-shadow: 10px 40px 50px 0px rgba(229, 233, 246, 0.4);
+  overflow: hidden;
+}
+
+.articleImage {
+  width: 100%;
+  height: 40%;
+}
+
+.articleCardGridContainer {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+  margin: 3% 0px;
+}
+
+.articleCardContent {
+  height: 53%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.articleCardText {
+  margin: 0px 33px;
+}
+
+.articleCardTitle {
+  color: #000;
+  font-family: "MulishRegular";
+  font-size: 21px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 32px;
+}
+
+.articleCardSubtitle {
+  color: #7d7987;
+  font-family: "MulishExtraLight";
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 300;
+  line-height: 28px;
+}
+
+.readMoreBtn {
+  margin: 0px 33px;
+  text-decoration: none;
+  color: #4089ed;
+  font-family: "MulishRegular";
+  font-size: 17px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 28px;
+}
+
+.readMoreArrow {
+  margin-inline: 2%;
+}
+
+.showMoreBtn {
+  margin-top: 6%;
+  width: 200px;
+  height: 56px;
+  border-radius: 55px;
+  background: #fff;
+  color: #458ff6;
+  font-family: "MulishRegular";
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 54px;
+  border: 1.4px solid #458ff6;
+}
