+++ src/components/SixSection/SixSection.module.css
@@ -1,0 +1,45 @@
+.sixSection {
+  justify-items: center;
+  display: inherit;
+  margin: 1% 5%;
+  margin-bottom: 15%;
+}
+
+.sixSectionTitle {
+  text-align: center;
+  color: #000;
+  font-family: "MulishRegular";
+  font-size: 36px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 56px; /* 155.556% */
+  margin-block: 0;
+}
+
+.backgroundImage {
+  position: absolute;
+  object-fit: cover;
+  z-index: -1;
+  overflow: hidden;
+}
+
+.sixSectionBGSky {
+  margin-left: 80%;
+  margin-top: 5%;
+  width: 36%;
+}
+
+.sixSectionBGMatrix {
+  left: 9%;
+  margin-top: 7%;
+  width: 10%;
+}
+
+.delimeter {
+  margin: 2% 0%;
+  width: 56px;
+  height: 2px;
+  flex-shrink: 0;
+  border-radius: 5px;
+  background: #000;
+}
