+++ src/components/ThirdSection/ThirdSection.module.css
@@ -1,0 +1,71 @@
+.thirdSection {
+  display: inline-flex;
+  color: #000;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  margin-bottom: 15%;
+}
+
+.thirdSectionContent {
+  width: 50%;
+  margin-left: 10em;
+}
+
+.delimeter {
+  margin: 2% 0%;
+  width: 56px;
+  height: 2px;
+  flex-shrink: 0;
+  border-radius: 5px;
+  background: #000;
+}
+
+.thirdSectionTitle {
+  color: #000;
+  font-family: "MulishRegular";
+  font-size: 48px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 56px; /* 116.667% */
+}
+
+.thirdSectionSubtitle {
+  color: #7d7987;
+  font-family: "MulishExtraLight";
+  font-size: 21px;
+  font-style: normal;
+  font-weight: 300;
+  line-height: 32px; /* 152.381% */
+}
+
+.thirdSectionImage {
+  width: 60%;
+}
+
+.thirdSectionBtn {
+  margin-top: 50px;
+  width: 200px;
+  height: 56px;
+  border-radius: 55px;
+  background: #fff;
+  color: #458ff6;
+  font-family: "MulishRegular";
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 54px;
+  border: 1.4px solid #458ff6;
+}
+
+.backgroundImage {
+  position: absolute;
+  object-fit: cover;
+  z-index: -1;
+  overflow: hidden;
+}
+.thirdSectionBGMatrix {
+  right: -10%;
+  width: 10%;
+  margin-top: 12%;
+}
