+++ src/components/ThirdSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ThirdSection";
